# backend/openapi.yaml
components:
  schemas:
...
    controller.apiResponse-any:
      properties:
        data: {}
        error:
          type: string
      type: object
...
    controller.apiResponse-array_model_Function:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.Function'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-array_model_OpDiv:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.OpDiv'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
    controller.apiResponse-array_model_Question:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.Question'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-model_Function:
      properties:
        data:
          $ref: '#/components/schemas/model.Function'
        error:
          type: string
      type: object
    controller.apiResponse-model_OpDiv:
      properties:
        data:
          $ref: '#/components/schemas/model.OpDiv'
        error:
          type: string
      type: object
    controller.apiResponse-model_Question:
      properties:
        data:
          $ref: '#/components/schemas/model.Question'
        error:
          type: string
      type: object
...
    model.Function:
      properties:
        datacenterenvironment:
          type: string
        description:
          type: string
        function:
          type: string
        functionid:
          type: integer
        order:
          type: integer
        pillarid:
          type: integer
        questionid:
          type: integer
      type: object
...
    model.OpDiv:
      properties:
        active:
          type: boolean
        code:
          type: string
        is_parent:
          description: |-
            IsParent and Active are pointers so an update can distinguish "not supplied"
            (nil, leave the column untouched) from an explicit false. With a non-pointer
            bool, any update that omits the field defaults to false: omitting is_parent
            would demote the HHS parent OpDiv (which drives HHS-wide authorization), and
            omitting active would silently deactivate an active OpDiv.
          type: boolean
        name:
          type: string
        opdiv_id:
          type: integer
      type: object
    model.Pillar:
      properties:
        order:
          type: integer
        pillar:
          type: string
        pillarid:
          type: integer
      type: object
...
    model.Question:
      properties:
        function:
          $ref: '#/components/schemas/model.Function'
        notesprompt:
          type: string
        order:
          type: integer
        pillar:
          $ref: '#/components/schemas/model.Pillar'
        pillarid:
          type: integer
        question:
          type: string
        questionid:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: API for the Zero Trust Maturity Framework (ZTMF).
  title: ZTMF API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /fismasystems/{fismasystemid}/questions:
    get:
      parameters:
      - description: FISMA system ID
        in: path
        name: fismasystemid
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_Question'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List questions relevant to a FISMA system
      tags:
      - fismasystems
...
  /functions:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_Function'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List all functions
      tags:
      - functions
    post:
      parameters:
      - description: Function ID
        in: path
        name: functionid
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.Function'
                description: Function to create or update
                summary: body
        description: Function to create or update
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Function'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a function
      tags:
      - functions
  /functions/{functionid}:
    get:
      parameters:
      - description: Function ID
        in: path
        name: functionid
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Function'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get a function by ID
      tags:
      - functions
    put:
      parameters:
      - description: Function ID
        in: path
        name: functionid
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.Function'
                description: Function to create or update
                summary: body
        description: Function to create or update
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Function'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a function
      tags:
      - functions
...
  /opdivs:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_OpDiv'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List all OpDivs
      tags:
      - opdivs
    post:
      parameters:
      - description: OpDiv ID
        in: path
        name: opdiv_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.OpDiv'
                description: OpDiv to create or update
                summary: body
        description: OpDiv to create or update
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_OpDiv'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update an OpDiv
      tags:
      - opdivs
  /opdivs/{opdiv_id}:
    put:
      parameters:
      - description: OpDiv ID
        in: path
        name: opdiv_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.OpDiv'
                description: OpDiv to create or update
                summary: body
        description: OpDiv to create or update
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_OpDiv'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update an OpDiv
      tags:
      - opdivs
  /questions:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_Question'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List all questions
      tags:
      - questions
    post:
      parameters:
      - description: Question ID
        in: path
        name: questionid
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.Question'
                description: Question to save
                summary: body
        description: Question to save
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Question'
          description: Created
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a question
      tags:
      - questions
  /questions/{questionid}:
    get:
      parameters:
      - description: Question ID
        in: path
        name: questionid
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Question'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get a question by ID
      tags:
      - questions
    put:
      parameters:
      - description: Question ID
        in: path
        name: questionid
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.Question'
                description: Question to save
                summary: body
        description: Question to save
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Question'
          description: Created
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a question
      tags:
      - questions
...
servers:
- url: /api/v1